+++ project/app/src/main/java/org/owntracks/android/net/mqtt/MQTTMessageProcessorEndpoint.kt
@@ -13,6 +13,7 @@ import com.fasterxml.jackson.databind.exc.InvalidFormatException
 import dagger.hilt.android.qualifiers.ApplicationContext
 import java.io.IOException
 import java.net.ConnectException
+import java.net.UnknownHostException
 import java.security.KeyStore
 import java.util.concurrent.ScheduledThreadPoolExecutor
 import java.util.stream.Collectors
@@ -227,7 +228,7 @@ class MQTTMessageProcessorEndpoint(
                               message.retained)
                           .also { Timber.v("MQTT message sent with messageId=${it.messageId}. ") }
                     }
-                    .apply { Timber.i("Message $message dispatched in $this") }
+                    .apply { Timber.i("Message ${message.messageId} sent in $this") }
               } catch (e: Exception) {
                 Timber.e(e, "Error publishing message $message")
                 when (e) {
@@ -423,6 +424,9 @@ class MQTTMessageProcessorEndpoint(
                           when (e.cause) {
                             is SSLHandshakeException ->
                                 Timber.e("$errorLog: ${(e.cause as SSLHandshakeException).message}")
+                            is UnknownHostException ->
+                                Timber.e(
+                                    "$errorLog: Unknown host \"${(e.cause as UnknownHostException).message}\"")
                             else -> Timber.e(e, errorLog)
                           }
                       REASON_CODE_SERVER_CONNECT_ERROR.toInt() -> {